+++ examples/filesharer/_request/request.sh
@@ -6,7 +6,7 @@ PUSH_FORMAT='{
     "receiver":"Bob",
     "req_data":{
         "method":"GET",
-        "host":"hidden-lake-service=filesharer",
+        "host":"hls-filesharer",
         "path":"/list?page=0"
     }
 }';