+++ examples/messenger/_request/main.go
@@ -21,7 +21,7 @@ const (
         "receiver":"%s",
         "req_data":{
 			"method":"POST",
-			"host":"hidden-lake-service=messenger",
+			"host":"hls-messenger",
 			"path":"/push",
 			"body":"%s"
 		}