+++ examples/remoter/_request/request.sh
@@ -8,7 +8,7 @@ PUSH_FORMAT='{
     "receiver":"Bob",
     "req_data":{
         "method":"POST",
-		"host":"hidden-lake-service=remoter",
+		"host":"hls-remoter",
 		"path":"/exec",
         "head":{
             "Password": "DpxJFjAlrs4HOWga0wk14mZqQSBo9DxK"